// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-bug-report-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminBugReportHttpResponse {
    attempt?: number;
    contentExpired?: boolean;
    id?: string;
    mergeRequestUrl?: string | null;
    receivedAtUtc?: string;
    sourceMessageId?: string;
    status?: string;
    subject?: string;
    summary?: string | null;
}