// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-bug-report-page-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminBugReportHttpResponse } from './admin-bug-report-http-response';

export interface AdminBugReportPageHttpResponse {
    isConfigured?: boolean;
    items?: Array<AdminBugReportHttpResponse>;
    totalItems?: number;
}